perf: Move route rules creation from middleware to plugin
